found a way to update the tag-map

+++ src/hvl/projectparmorel/reward/PreferRepairingHighInContextHierarchyPreference.java
@@ -5,13 +5,13 @@ import hvl.projectparmorel.general.Error;
 
 public class PreferRepairingHighInContextHierarchyPreference extends Preference {
 	
-	public PreferRepairingHighInContextHierarchyPreference(double weight) {
+	public PreferRepairingHighInContextHierarchyPreference(int weight) {
 		super(weight);
 	}
 	
 	@Override
-	public double rewardActionForError(Error currentErrorToFix, Action action) {
-		double reward = 0;
+	public int rewardActionForError(Error currentErrorToFix, Action action) {
+		int reward = 0;
 		
 //		int contextId = action.getHierarchy();
 		if (action.getHierarchy() == 1) {
